fix: Refactored availability tests for less flakiness
